Move out `QuestionMoveToast`
